+++ apps/macos/LauncherApp/look-app/Window/WindowConfigurator.swift
@@ -18,19 +18,24 @@ struct WindowConfigurator: NSViewRepresentable {
 
     private func configureWindow(from view: NSView) {
         guard let window = view.window else { return }
-        window.styleMask.remove(.titled)
+
+        window.styleMask.insert(.titled)
         window.styleMask.remove(.closable)
         window.styleMask.remove(.miniaturizable)
         window.styleMask.remove(.resizable)
-        window.styleMask.insert(.borderless)
+        window.styleMask.remove(.borderless)
         window.styleMask.insert(.fullSizeContentView)
         window.titleVisibility = .hidden
         window.titlebarAppearsTransparent = true
+        window.titlebarSeparatorStyle = .none
         window.toolbar = nil
         window.isOpaque = false
         window.backgroundColor = .clear
         window.isMovableByWindowBackground = true
         window.hasShadow = true
+        window.setContentBorderThickness(0, for: .maxY)
+        window.collectionBehavior.insert(.moveToActiveSpace)
+        window.collectionBehavior.insert(.fullScreenAuxiliary)
         window.standardWindowButton(.closeButton)?.isHidden = true
         window.standardWindowButton(.miniaturizeButton)?.isHidden = true
         window.standardWindowButton(.zoomButton)?.isHidden = true
